+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
  - Add £5/mo dividend goal.
  - Add a howto to the holdings page.
+ - Show pie average yield and pie weighted yield on forcasting (index) page.
 ### Changed
 
  - Improve visuals of the tabs.